lingkaran.py

+++ lingkaran.py
@@ -1,4 +1,4 @@
 r = int(input("Masukkan Jari - jari : "))
-luas = float(3.14 * r * r)
+luas = 3.14 * r * r
 
 print(f'Hasilnya adalah : {luas:.2f}')
